merge Navs
We introduce the flag parameter antipattern on purpose here.

We want the specificity of each form to disapear in the long run so we
have one form to rule them all.

This pattern should encourage developer not to add to many specificity
in forms.

see. https://martinfowler.com/bliki/FlagArgument.html