feat: update contact_technique prefill
Mind that this only work by accident as team_members array is merged
recursively with the one in demarches.json on demarche change. This
means that the first item of the team_members arrays are merged together.
It work only because contact_technique is added at the first place in
the array, hence it is saved first as a team_member, hence updating the
demarche will update the first item of the array, which luckily is the
good one.